// src/app/official-demo/_enums/demo-form-field.enum.ts
/** Chaves do `FormGroup` da seção Campos na demo. */
export enum DemoFormField {
  Texto = 'texto',
  TextoRo = 'textoRo',
  Senha = 'senha',
  SenhaRo = 'senhaRo',
  Observacao = 'observacao',
  ObservacaoRo = 'observacaoRo',
  Inteiro = 'inteiro',
  InteiroRo = 'inteiroRo',
  DecimalBr = 'decimalBr',
  DecimalBrRo = 'decimalBrRo',
  Data = 'data',
  DataRo = 'dataRo',
  Cep = 'cep',
  CepRo = 'cepRo',
  Telefone = 'telefone',
  TelefoneRo = 'telefoneRo',
  TelefoneFixo = 'telefoneFixo',
  TelefoneFixoRo = 'telefoneFixoRo',
  CpfCnpj = 'cpfCnpj',
  CpfCnpjRo = 'cpfCnpjRo',
  Estado = 'estado',
  EstadoRo = 'estadoRo',
  EstadoBusca = 'estadoBusca',
  EstadoBuscaRo = 'estadoBuscaRo',
  EstadosMulti = 'estadosMulti',
  EstadosMultiRo = 'estadosMultiRo',
  Aceite = 'aceite',
  AceiteRo = 'aceiteRo',
  Ativo = 'ativo',
  AtivoRo = 'ativoRo',
}